+++ Assets/Scripts/StaticConfig/Encounters/EnemySpawnVariables.cs
@@ -3,6 +3,6 @@
   /// Encapsulates the variables that decide an enemy's spawn chance
   /// </summary>
   public class EnemySpawnVariables {
-    public int DifficultyRating { get; set; }
+    public float DifficultyRating { get; set; }
   }
 }
